requirements.txt
requests>=2.31.0
cryptography>=41.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
rich>=13.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
